+++ src/.dev.env
@@ -5,3 +5,5 @@ DB_HOST=localhost
 DB_PORT=5432
 DB_USER=admin
 DB_PASSWORD=password
+OPENCAGE_API_KET=b8fcd228862e4ee9b8efd5817fac44ea
+GOOGLE_MAPS_API_KEY=AIzaSyDU48f3kKemLgrFn78M1Rd_kV9kTH34XvA
