--- src/coordinates/tests/test_model.py
@@ -1,7 +1,0 @@
-from coordinates.service.calculations import coordinates_distance
-from rest_framework.test import APITestCase
-
-
-class TestTrue(APITestCase):
-    def test_true(self):
-        assert 1 == 1
